backend/requirements.txt
fastapi
uvicorn
pytest
celery
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
email-validator
python-jose
bcrypt
qdrant-client
redis
slowapi
limits
tree-sitter==0.23.0
tree-sitter-python==0.23.0
ollama
groq
sentence-transformers
httpx>=0.27,<0.28
